feat: 적금 가입 기능 구현

+++ seasonthon_team_25_fe/lib/feature/bank/saving/domain/entities/savings_maturity_preview_entity.dart
@@ -16,7 +16,7 @@ abstract class SavingsMaturityInfoEntity with _$SavingsMaturityInfoEntity {
     required int principal,
     required int interest,
     required int tax,
-    required double totalAmount,
+    required int totalAmount,
     required double interestRate,
   }) = _SavingsMaturityInfoEntity;
 }
